requirements.txt
flask==3.0.0
python-dotenv==1.0.0
websockets==12.0
pyaudio==0.2.14
assemblyai==0.20.0
python-socketio==5.11.0
flask-socketio==5.3.6 